requirements.txt
kafka-python-ng
pyspark==3.5.0
pymongo==4.6.0
elasticsearch==8.10.0
minio==7.2.0
pandas
numpy
pyarrow
requests==2.31.0
apache-airflow==2.7.0
apache-airflow-providers-apache-spark==4.7.0
streamlit==1.28.0
folium==0.14.0
streamlit-folium==0.17.0
python-dotenv==1.0.0
pytz==2023.3